architecture: move soccer data to lib/quiz/data; refactor useSoccerGame to use factory (#652)
- Move SOCCER_QUESTIONS / SOCCER_CATEGORIES / SoccerQuestion to
  lib/quiz/data/soccer.ts (canonical location next to other quiz data)
- app/games/soccer/data/soccer.ts becomes a re-export shim so existing
  imports continue to resolve
- soccerGameStore.ts simplified to only manage showGoal (question /
  category management moved to createCategoryIndexQuizHook)
- lib/quiz/useSoccerGame.ts rewritten: calls createCategoryIndexQuizHook
  factory, then wraps selectAnswer to inject goal-celebration animation
