+++ gamesformykids/lib/quiz/data/soccer.ts
@@ -1,0 +1,142 @@
+// ⚽ משחק כדורגל — טריוויה ושאלות
+export type SoccerQuestion = {
+  id: number;
+  question: string;
+  answers: [string, string, string, string];
+  correctIndex: number;
+  emoji: string;
+  category: 'כללי' | 'שחקנים' | 'קבוצות' | 'חוקים' | 'טכניקה';
+  funFact: string;
+};
+
+export const SOCCER_QUESTIONS: SoccerQuestion[] = [
+  {
+    id: 1, emoji: '⚽', category: 'כללי',
+    question: 'כמה שחקנים יש לכל קבוצה בכדורגל?',
+    answers: ['9', '10', '11', '12'],
+    correctIndex: 2,
+    funFact: '11 שחקנים לכל קבוצה — כולל שוער אחד!'
+  },
+  {
+    id: 2, emoji: '🥅', category: 'חוקים',
+    question: 'כמה דקות נמשכת משחק כדורגל רגיל?',
+    answers: ['60', '80', '90', '120'],
+    correctIndex: 2,
+    funFact: 'משחק כדורגל מחולק לשני מחצות של 45 דקות כל אחת!'
+  },
+  {
+    id: 3, emoji: '🏆', category: 'כללי',
+    question: 'איזה כלי משמש לשמור על שער בכדורגל?',
+    answers: ['מקל', 'כפפות', 'קסדה', 'חגורה'],
+    correctIndex: 1,
+    funFact: 'השוערים לובשים כפפות מיוחדות שעוזרות לתפוס את הכדור!'
+  },
+  {
+    id: 4, emoji: '🌍', category: 'כללי',
+    question: 'כיצד קוראים לתחרות הכדורגל הגדולה בעולם?',
+    answers: ['אליפות אירופה', 'גביע העולם', 'ליגת האלופות', 'סופרבול'],
+    correctIndex: 1,
+    funFact: 'גביע העולם בכדורגל מתקיים כל 4 שנים ומשתתפות בו 32 נבחרות!'
+  },
+  {
+    id: 5, emoji: '🇧🇷', category: 'קבוצות',
+    question: 'איזו נבחרת זכתה הכי הרבה פעמים בגביע העולם?',
+    answers: ['גרמניה', 'איטליה', 'ברזיל', 'ארגנטינה'],
+    correctIndex: 2,
+    funFact: 'ברזיל זכתה 5 פעמים בגביע העולם — יותר מכל מדינה אחרת!'
+  },
+  {
+    id: 6, emoji: '⭐', category: 'שחקנים',
+    question: 'מי נקרא "El Pulga" (הפרעוש) בכדורגל?',
+    answers: ['רונאלדו', 'מסי', 'נמאר', 'מבאפה'],
+    correctIndex: 1,
+    funFact: 'ליאונל מסי מכונה ה"פרעוש" בגלל קומתו הנמוכה ומהירות הריצה שלו!'
+  },
+  {
+    id: 7, emoji: '🚩', category: 'חוקים',
+    question: 'מה מנפנף השופט הצדדי כשכדור יוצא מהמגרש?',
+    answers: ['כרטיס אדום', 'דגל', 'שריקה', 'מחבט'],
+    correctIndex: 1,
+    funFact: 'השופטים הצדדיים מנופפים דגל כדי לאותת לשופט הראשי!'
+  },
+  {
+    id: 8, emoji: '🟨', category: 'חוקים',
+    question: 'מה קורה לשחקן שמקבל שני כרטיסים צהובים?',
+    answers: ['קנס כספי', 'יוצא מהמגרש', 'בעיטת עונשין לקבוצה השנייה', 'ממשיך לשחק'],
+    correctIndex: 1,
+    funFact: 'שני כרטיסים צהובים = כרטיס אדום, והשחקן מורחק מהמשחק!'
+  },
+  {
+    id: 9, emoji: '🏟️', category: 'כללי',
+    question: 'מה השם של המגרש שעליו משחקים כדורגל?',
+    answers: ['שדה', 'רחוב', 'מגרש', 'אצטדיון'],
+    correctIndex: 3,
+    funFact: 'כינוי נוסף: "קופסה" — כי הצלחת הגבולות יוצרות צורה מלבנית!'
+  },
+  {
+    id: 10, emoji: '🎽', category: 'כללי',
+    question: 'מי בקבוצה לובש חולצה בצבע שונה מכל שאר השחקנים?',
+    answers: ['שחקן ה-10', 'הקפטן', 'השוער', 'החליף'],
+    correctIndex: 2,
+    funFact: 'השוער לובש חולצה בצבע שונה כדי שיהיה ניכר לשחקנים ולשופטים!'
+  },
+  {
+    id: 11, emoji: '⚽', category: 'טכניקה',
+    question: 'כיצד קוראים לבעיטה שנועדת לעבור מעל הרגל של היריב?',
+    answers: ['חיתוך', 'לופ', 'נוגה', 'שיחור'],
+    correctIndex: 1,
+    funFact: 'הלופ הוא תרגיל טכני שבו הכדור עובר מעל הרגל היריב!'
+  },
+  {
+    id: 12, emoji: '🥇', category: 'שחקנים',
+    question: 'איזה פרס נותנים לשחקן הטוב ביותר בעולם כל שנה?',
+    answers: ['גביע הזהב', "כדור הזהב", 'נעל הזהב', 'תעודת הצטיינות'],
+    correctIndex: 1,
+    funFact: "כדור הזהב הוא הפרס היוקרתי ביותר לשחקן כדורגל בעולם!"
+  },
+  {
+    id: 13, emoji: '🔴', category: 'חוקים',
+    question: 'מה אורכו של מגרש כדורגל סטנדרטי?',
+    answers: ['50-60 מטר', '70-80 מטר', '90-120 מטר', '150-200 מטר'],
+    correctIndex: 2,
+    funFact: 'מגרש כדורגל סטנדרטי הוא בין 90 ל-120 מטר אורך!'
+  },
+  {
+    id: 14, emoji: '🇦🇷', category: 'שחקנים',
+    question: 'מאיזה מדינה בא ליאונל מסי?',
+    answers: ['ברזיל', 'ספרד', 'אורוגוואי', 'ארגנטינה'],
+    correctIndex: 3,
+    funFact: 'מסי נולד ב-1987 בעיר רוסאריו, ארגנטינה!'
+  },
+  {
+    id: 15, emoji: '👟', category: 'טכניקה',
+    question: 'מה שמו של הנעל המיוחדת לכדורגל?',
+    answers: ['סניקרס', 'קריסי', "סוליות", 'קינגס'],
+    correctIndex: 2,
+    funFact: 'נעלי כדורגל עם סוליות (ספייקס) מספקות אחיזה טובה על הדשא!'
+  },
+  {
+    id: 16, emoji: '🤲', category: 'חוקים',
+    question: 'מה עושים כששחקן נפגע ונופל במגרש?',
+    answers: ['ממשיכים לשחק', 'השחקן יוצא', 'עצירת המשחק', 'כרטיס צהוב'],
+    correctIndex: 2,
+    funFact: 'כשיש פציעה, השופט מצייץ לעצור את המשחק לטיפול!'
+  },
+  {
+    id: 17, emoji: '⚽', category: 'קבוצות',
+    question: 'מהי הקבוצה הנקראת "הקסרים הלבנים"?',
+    answers: ['ברצלונה', 'ריאל מדריד', 'מנצ\'סטר יונייטד', 'בייארן מינכן'],
+    correctIndex: 1,
+    funFact: 'ריאל מדריד מכונה "הקסרים הלבנים" בגלל חולצתם הלבנה המיוחדת!'
+  },
+  {
+    id: 18, emoji: '🏅', category: 'כללי',
+    question: 'מה קורה אם תוצאת המשחק שווה בסוף 90 דקות?',
+    answers: ['הפסד לשתי הקבוצות', 'תוספת זמן', 'קבוצה ביתית מנצחת', 'חוזרים ביום אחר'],
+    correctIndex: 1,
+    funFact: 'אם התוצאה לא מוכרעת, משחקים 30 דקות נוספות ואם לא — בעיטות עונשין!'
+  },
+];
+
+export const SOCCER_CATEGORIES = ['הכל', 'כללי', 'שחקנים', 'קבוצות', 'חוקים', 'טכניקה'] as const;
+export type SoccerCategory = typeof SOCCER_CATEGORIES[number];
